+++ src/Repository/ApiKeyRepository.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Repository;
+
+use App\Entity\ApiKey;
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Persistence\ManagerRegistry;
+
+/**
+ * @method null|ApiKey find($id, $lockMode = null, $lockVersion = null)
+ * @method null|ApiKey findOneBy(array $criteria, array $orderBy = null)
+ * @method ApiKey[]    findAll()
+ * @method ApiKey[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class ApiKeyRepository extends ServiceEntityRepository
+{
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, ApiKey::class);
+    }
+
+    public function findOneByKeyHash(string $keyHash): ?ApiKey
+    {
+        return $this->findOneBy(['keyHash' => $keyHash], null);
+    }
+}